+++ tests/web/registry/test_signature_import_from_raw_text_signature.py
