+++ tests/web/utility/test_is_valid_contract_abi.py
