+++ tox.ini
@@ -4,7 +4,7 @@ exclude=**/migrations/*
 
 [tox]
 envlist=
-    py35-django19,
+    py35-{django19,contracts},
     flake8
 skipsdist=True
 
@@ -12,15 +12,19 @@ skipsdist=True
 basepython=python3.5
 commands=
     py.test tests/
+commands=
+    django19: py.test {posargs:tests/web}
+    contracts: py.test {posargs:tests/contracts}
 passenv=
     HUEY_REDIS_HOST
     HUEY_REDIS_PORT
+    DATABASE_URL
+    SOLC_BINARY
 setenv=
     DJANGO_SETTINGS_MODULE=func_sig_registry.settings
     DJANGO_SECRET_KEY=not-a-real-secret-key
     DJANGO_SECURE_SSL_REDIRECT=False
     DJANGO_DEBUG=True
-    DATABASE_URL=sqlite://:memory:
 deps =
     -r{toxinidir}/requirements.txt
     -r{toxinidir}/requirements-dev.txt
